Moved project and update files to the updater project
