+++ src/Sidekick.Common.Updater/Components/Update.fr.resx
