+++ src/Sidekick.Common.Updater/Components/Update.ko.resx
