+++ src/Sidekick.Common.Updater/Components/Update.razor
