+++ src/Sidekick.Common.Updater/Components/Update.resx
