+++ src/Sidekick.Common.Updater/Components/UpdateBanner.razor
