+++ src/Sidekick.Common.Updater/Sidekick.Common.Updater.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk.Razor">
 
     <PropertyGroup>
         <TargetFramework>net8.0</TargetFramework>
@@ -10,4 +10,14 @@
       <PackageReference Include="Velopack" Version="0.0.942" />
     </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Sidekick.Common.Ui\Sidekick.Common.Ui.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Content Include="..\Sidekick.Common.Ui\_Imports.razor">
+      <Link>_Imports.razor</Link>
+    </Content>
+  </ItemGroup>
+  
 </Project>
