+++ src/Sidekick.Common.Updater/StartupExtensions.cs
@@ -7,6 +7,7 @@ public static class StartupExtensions
     public static IServiceCollection AddSidekickUpdater(this IServiceCollection services)
     {
         services.AddTransient<IAutoUpdater, AutoUpdater>();
+        services.AddSidekickModule(typeof(StartupExtensions).Assembly);
 
         return services;
     }
